requirements.txt
streamlit==0.60.0
numpy==1.17.2
transformers==2.10.0
newspaper3k==0.2.8
torch==1.4.0
nltk==3.5